--- DisguiseUnityRenderStream/Runtime/TemporaryTexture2DManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fd9763cf8fae5e54dbbe86d826f180c3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
